stopt nu als je geen knop indrukt

+++ src/walkgame/objects/Player.java
@@ -9,17 +9,17 @@ public class Player extends Character implements Controllable
     private static final int PLAYER_HEALTH = 100;
     private static final int PLAYER_SPEED = 2;
 
-    public Player(float x, float y, Image image, String name, int health, float speed)
+    public Player(double x, double y, Image image, String name, int health, float speed)
     {
         super(x, y, image, name, health, speed);
     }
 
-    public Player(float x, float y, Image image, String name)
+    public Player(double x, double y, Image image, String name)
     {
         this(x, y, image, name, PLAYER_HEALTH, PLAYER_SPEED);
     }
 
-    public Player(float x, float y, String name)
+    public Player(double x, double y, String name)
     {
         this(x, y, new Image("walkgame/res/playerSouth.jpg"), name);
     }
